server/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-sqlalchemy
flask-mail
psycopg2-binary
python-dotenv
werkzeug
cloudinary
